[update] jsb - 결제 타입 BigDemical -> Long 전체 변경 완료

+++ src/main/java/com/sashimi/cart/application/port/CourseInfo.java
@@ -1,12 +1,10 @@
 package com.sashimi.cart.application.port;
 
-import java.math.BigDecimal;
-
 public record CourseInfo(
 
         Long courseId,
         String title,
-        BigDecimal price,
+        Long price,
         String thumbnail,
         String instructorName,
         boolean purchasable
